# Reuse empty value store arrays.