--- a/tests/Avalonia.Benchmarks/NullCursorFactory.cs
+++ b/tests/Avalonia.Benchmarks/NullCursorFactory.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Avalonia.Input;
+using Avalonia.Platform;
+
+namespace Avalonia.Benchmarks
+{
+    internal class NullCursorFactory : IStandardCursorFactory
+    {
+        public IPlatformHandle GetCursor(StandardCursorType cursorType)
+        {
+            return new PlatformHandle(IntPtr.Zero, "null");
+        }
+    }
+}