## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Removes one extra (tiny) allocation from each `AvaloniaObject` instance by simply reusing empty array.
No big gains but this change is pretty much free (and removes like 10k array allocations from ControlCatalog startup).

Also I've noticed that we were still using Moq cursor factory in benchmarks which is kinda slow since it uses reflection.

## Benchmarks

Before
|         Method |        Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |   Gen 2 |  Allocated |
|--------------- |------------:|----------:|----------:|---------:|---------:|--------:|-----------:|
| CreateCalendar | 22,691.6 us | 159.25 us | 141.17 us | 718.7500 | 343.7500 | 31.2500 | 4240.21 KB |
|   CreateButton |    177.1 us |   0.81 us |   0.64 us |   9.7656 |   0.9766 |       - |   40.28 KB |
|  CreateTextBox |  1,268.1 us |   6.48 us |   6.07 us |  52.7344 |  27.3438 |  3.9063 |  213.33 KB |

After
         Method |        Mean |     Error |   StdDev |    Gen 0 |    Gen 1 |  Gen 2 |  Allocated |
|--------------- |------------:|----------:|---------:|---------:|---------:|-------:|-----------:|
| CreateCalendar | 22,431.0 us | 120.56 us | 94.12 us | 687.5000 | 312.5000 |      - | 4220.51 KB |
|   CreateButton |    178.0 us |   1.29 us |  1.21 us |   9.7656 |   0.9766 |      - |    40.2 KB |
|  CreateTextBox |  1,267.6 us |  12.46 us | 11.66 us |  52.7344 |  27.3438 | 3.9063 |  212.29 KB |